website-backend/requirements.txt
fastapi
uvicorn[standard]
aiosmtplib
slowapi
python-dotenv
pydantic[email]
httpx
